// __generated__/graph-types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
};

export type ChangePasswordRequest = {
  email: Scalars['String'];
  newPassword: Scalars['String'];
  password: Scalars['String'];
};

export type CreateMovieRequest = {
  description: Scalars['String'];
  directorName: Scalars['String'];
  movieName: Scalars['String'];
  releaseDate: Scalars['Date'];
};

export type CreateReviewRequest = {
  comment: Scalars['String'];
  movieId: Scalars['Int'];
  rating: Scalars['Int'];
  userId: Scalars['Int'];
};
// ...
export type MovieResponse = {
  __typename?: 'MovieResponse';
  description: Scalars['String'];
  directorName: Scalars['String'];
  id: Scalars['Int'];
  movieName: Scalars['String'];
  releaseDate: Scalars['Date'];
  user?: Maybe<UserResponse>;
};
// ...
export type MutationCreateReviewArgs = {
  review?: InputMaybe<CreateReviewRequest>;
};


export type MutationDeleteMovieArgs = {
  id: Scalars['Int'];
};


export type MutationDeleteReviewArgs = {
  id: Scalars['Int'];
};


export type MutationSignupArgs = {
  user: UserSignup;
};


export type MutationUpdateMovieArgs = {
  id: Scalars['Int'];
  movie: UpdateMovieRequest;
};


export type MutationUpdateReviewArgs = {
  id: Scalars['Int'];
  review: UpdateReviewRequest;
};

export type PaginatedMovieList = {
  __typename?: 'PaginatedMovieList';
  meta: PaginationResponse;
  movies: Array<Maybe<MovieResponse>>;
};

export type PaginatedReviewList = {
  __typename?: 'PaginatedReviewList';
  meta: PaginationResponse;
  reviews: Array<Maybe<ReviewResponse>>;
};

export type PaginationFields = {
  pageNumber?: InputMaybe<Scalars['Int']>;
  pageSize?: InputMaybe<Scalars['Int']>;
};

export type PaginationResponse = {
  __typename?: 'PaginationResponse';
  pageSize: Scalars['Int'];
  totalPages: Scalars['Int'];
};
// ...
export type QueryListReviewsArgs = {
  filter?: InputMaybe<ReviewFilters>;
  movieId: Scalars['Int'];
  pagination?: InputMaybe<PaginationFields>;
  sort?: InputMaybe<ReviewSortFields>;
};


export type QueryLoginArgs = {
  credentials: LoginCredentials;
};

export type ReviewFilters = {
  comment?: InputMaybe<Scalars['String']>;
  rating?: InputMaybe<Scalars['Int']>;
};

export type ReviewResponse = {
  __typename?: 'ReviewResponse';
  comment: Scalars['String'];
  id: Scalars['Int'];
  movie?: Maybe<MovieResponse>;
  rating: Scalars['Int'];
  user?: Maybe<UserResponse>;
};

export type ReviewSortFields = {
  field: SortingReviewFields;
  isAscending: Scalars['Boolean'];
};

export enum SortingMovieFields {
  DirectorName = 'directorName',
  Id = 'id',
  MovieName = 'movieName',
  ReleaseDate = 'releaseDate'
}

export enum SortingReviewFields {
  Id = 'id',
  Rating = 'rating'
}

export type UpdateMovieRequest = {
  description?: InputMaybe<Scalars['String']>;
  directorName?: InputMaybe<Scalars['String']>;
  movieName?: InputMaybe<Scalars['String']>;
  releaseDate?: InputMaybe<Scalars['Date']>;
};

export type UpdateReviewRequest = {
  comment?: InputMaybe<Scalars['String']>;
  rating?: InputMaybe<Scalars['Int']>;
};

export type User = {
  __typename?: 'User';
  email: Scalars['String'];
  username: Scalars['String'];
};

export type UserResponse = {
  __typename?: 'UserResponse';
  email: Scalars['String'];
  id: Scalars['Int'];
  username: Scalars['String'];
};

export type UserSignup = {
  email: Scalars['String'];
  password: Scalars['String'];
  username: Scalars['String'];
};

export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;

export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;